requirements.txt
docopt~=0.6.2
mariadb~=1.0.6
bs4~=0.0.1
beautifulsoup4~=4.9.3
requests~=2.25.1
setuptools~=54.1.1
Flask~=1.1.2
dearpygui~=0.6.281
pandas~=1.2.3
plotly~=4.14.3
python-dateutil~=2.8.1